HHH-11147 - Allow enhanced entities to be returned in a completely uninitialized state  - post Andrea test fixes: javadoc, comments, etc
