+++ hibernate-core/src/main/java/org/hibernate/event/internal/DefaultMergeEventListener.java
@@ -14,7 +14,6 @@ import org.hibernate.HibernateException;
 import org.hibernate.ObjectDeletedException;
 import org.hibernate.StaleObjectStateException;
 import org.hibernate.WrongClassException;
-import org.hibernate.bytecode.enhance.spi.interceptor.EnhancementAsProxyLazinessInterceptor;
 import org.hibernate.engine.internal.Cascade;
 import org.hibernate.engine.internal.CascadePoint;
 import org.hibernate.engine.spi.CascadingAction;
@@ -209,40 +208,54 @@ public class DefaultMergeEventListener extends AbstractSaveEventListener impleme
 		LOG.trace( "Merging transient instance" );
 
 		final Object entity = event.getEntity();
-		final EventSource source = event.getSession();
+		if ( entity instanceof PersistentAttributeInterceptable ) {
+
+		}
+
+		final EventSource session = event.getSession();
 
 		final String entityName = event.getEntityName();
-		final EntityPersister persister = source.getEntityPersister( entityName, entity );
+		final EntityPersister persister = session.getEntityPersister( entityName, entity );
 
-		final Serializable id = persister.hasIdentifierProperty() ?
-				persister.getIdentifier( entity, source ) :
-				null;
-		if ( copyCache.containsKey( entity ) ) {
-			persister.setIdentifier( copyCache.get( entity ), id, source );
+		final Serializable id = persister.hasIdentifierProperty()
+				? persister.getIdentifier( entity, session )
+				: null;
+
+		final Object copy;
+		final Object existingCopy = copyCache.get( entity );
+		if ( existingCopy != null ) {
+			persister.setIdentifier( copyCache.get( entity ), id, session );
+			copy = existingCopy;
 		}
 		else {
-			final Object copy = source.instantiate( persister, id );
+			copy = session.instantiate( persister, id );
 
 			//before cascade!
 			( (MergeContext) copyCache ).put( entity, copy, true );
 		}
 
-		final Object copy = copyCache.get( entity );
-
 		// cascade first, so that all unsaved objects get their
 		// copy created before we actually copy
 		//cascadeOnMerge(event, persister, entity, copyCache, Cascades.CASCADE_BEFORE_MERGE);
-		super.cascadeBeforeSave( source, persister, entity, copyCache );
-		copyValues( persister, entity, copy, source, copyCache, ForeignKeyDirection.FROM_PARENT );
+		super.cascadeBeforeSave( session, persister, entity, copyCache );
+		copyValues( persister, entity, copy, session, copyCache, ForeignKeyDirection.FROM_PARENT );
 
-		saveTransientEntity( copy, entityName, event.getRequestedId(), source, copyCache );
+		saveTransientEntity( copy, entityName, event.getRequestedId(), session, copyCache );
 
 		// cascade first, so that all unsaved objects get their
 		// copy created before we actually copy
-		super.cascadeAfterSave( source, persister, entity, copyCache );
-		copyValues( persister, entity, copy, source, copyCache, ForeignKeyDirection.TO_PARENT );
+		super.cascadeAfterSave( session, persister, entity, copyCache );
+		copyValues( persister, entity, copy, session, copyCache, ForeignKeyDirection.TO_PARENT );
 
 		event.setResult( copy );
+
+		if ( copy instanceof PersistentAttributeInterceptable ) {
+			final PersistentAttributeInterceptable interceptable = (PersistentAttributeInterceptable) copy;
+			final PersistentAttributeInterceptor interceptor = interceptable.$$_hibernate_getInterceptor();
+			if ( interceptor == null ) {
+				persister.getInstrumentationMetadata().injectInterceptor( copy, id, session );
+			}
+		}
 	}
 
 	private void saveTransientEntity(
